Move buildScenario to to preset too
